requirements.txt
neat-python==0.92
pygame==2.6.1
graphviz==0.20.3